get recommended assessments
